fix(trainer): print task-group summary from the async buffer path too
The colorful summary from #784 lived only in AgentFlowEngine.execute_tasks
(the sync/eval path). Async training dispatches rollouts through the buffer/
coordinator instead, so groups finished silently — the per-episode line still
printed (it's in process_task_with_retry) but no group summary appeared.

Extract the formatter into rllm/utils/group_summary.py (dependency-light, no
engine↔trainer cycle) and call it from TrajectoryGroupBuffer's per-group
completion hook, passing the buffer verdict (queued vs. filtered:<reason>) that
only that path knows. execute_tasks now imports the same shared formatter.
